Added note about test coverage.
The script that *runs* the simulation is not testes, as well, the
game model is not tested to be fair as there's no real way to do
that (unless the prizes are exposed before the picks are made).
The test could run it enough times and complain if there are no
wins/losses).